Task_2/requirements.txt
pandas==1.5.3
mysql-connector-python==8.0.30
python-dotenv==0.21.0
numpy==1.23.5  
